convert stick_test_data to class
